requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==0.19.0
mistralai==0.0.7
flask-cors==3.0.10 